Cleanup: Separate Builtins from direct evaluator, make evaluator naming consistent
