+++ pact-tng.cabal
@@ -270,6 +270,7 @@ library
     Pact.Core.IR.Desugar
 
     Pact.Core.IR.Eval.Direct.Evaluator
+    Pact.Core.IR.Eval.Direct.CoreBuiltin
     Pact.Core.IR.Eval.Direct.ReplBuiltin
     Pact.Core.IR.Eval.Direct.Types
 
@@ -277,10 +278,10 @@ library
     Pact.Core.IR.Eval.Runtime
     Pact.Core.IR.Eval.Runtime.Types
     Pact.Core.IR.Eval.Runtime.Utils
-    Pact.Core.IR.Eval.CEK
+    Pact.Core.IR.Eval.CEK.Evaluator
     Pact.Core.IR.Eval.CEK.Types
     Pact.Core.IR.Eval.CEK.Utils
-    Pact.Core.IR.Eval.CoreBuiltin
+    Pact.Core.IR.Eval.CEK.CoreBuiltin
     Pact.Core.IR.ModuleHashing
     Pact.Core.IR.ConstEval
